2.0: bunch of adaptations to GUI
A lot of sprites were removed, I tried to find similar looking, but
sometimes it was not possible at all, when for example _hover or _black
version of some sprite was removed.

Same thing happened to a lot of very specific GUI themes, in those cases
I just replaced it with more general forms.